+++ src/main/java/exception/DuplicateNameException.java
@@ -1,7 +1,5 @@
 package exception;
 
-import view.input.ErrorMessage;
-
 public class DuplicateNameException extends IllegalArgumentException {
 
     public DuplicateNameException() {
