+++ src/main/java/exception/EmptyInputException.java
@@ -1,7 +1,5 @@
 package exception;
 
-import view.input.ErrorMessage;
-
 public class EmptyInputException extends IllegalArgumentException {
 
     public EmptyInputException() {
