+++ src/main/java/exception/InvalidLineWeightException.java
@@ -1,7 +1,5 @@
 package exception;
 
-import view.input.ErrorMessage;
-
 public class InvalidLineWeightException extends IllegalArgumentException {
 
     public InvalidLineWeightException() {
