+++ src/main/java/exception/InvalidResultsCount.java
@@ -1,7 +1,5 @@
 package exception;
 
-import view.input.ErrorMessage;
-
 public class InvalidResultsCount extends IllegalArgumentException {
 
     public InvalidResultsCount() {
