+++ src/test/java/controller/GameResultControllerTest.java
@@ -8,7 +8,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import view.input.ErrorMessage;
+import exception.ErrorMessage;
 import view.input.InputView;
 import view.output.OutputView;
